refactor: add UserController
